Adding movie validators and responses

+++ src/main/java/com/hackathon/equipo2/CineHackacthon/services/MovieService.java
@@ -2,7 +2,10 @@ package com.hackathon.equipo2.CineHackacthon.services;
 
 import com.hackathon.equipo2.CineHackacthon.models.MovieModel;
 import com.hackathon.equipo2.CineHackacthon.repositories.MovieRepository;
+import com.hackathon.equipo2.CineHackacthon.services.responses.MovieServiceResponse;
+import com.hackathon.equipo2.CineHackacthon.validators.MovieValidator;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,26 +15,43 @@ public class MovieService {
 
     @Autowired
     MovieRepository movieRepository;
-    public List<MovieModel> findAll() {
 
-        return this.movieRepository.findAll();
+    @Autowired
+    MovieValidator movieValidator;
+
+
+    public MovieServiceResponse findAll() {
+        return movieValidator.existMovieList(this.movieRepository.findAll());
     }
 
-    public MovieModel findById(Long id) {
+    public MovieServiceResponse findById(Long id) {
+        MovieServiceResponse response = movieValidator.exist(this.movieRepository.findById(id));
+        return response;
 
-
-        return this.movieRepository.findById(id);
     }
 
-    public Boolean add(MovieModel movie) {
-        return this.movieRepository.add(movie);
+    public MovieServiceResponse add(MovieModel movie) {
+        MovieServiceResponse response = movieValidator.createMovie(this.movieRepository.findById(movie.getId()));
+        if(response.getHttpStatus() == HttpStatus.OK){
+            this.movieRepository.add(movie);
+        }
+        return response;
+
     }
 
-    public Boolean remove(Long id) {
-        return this.movieRepository.remove(id);
+    public MovieServiceResponse remove(Long id) {
+        MovieServiceResponse response = movieValidator.removeMovie(this.movieRepository.findById(id));
+        if(response.getHttpStatus() == HttpStatus.OK) {
+             this.movieRepository.remove(id);
+        }
+        return response;
     }
 
-    public Boolean update(MovieModel movieModel) {
-        return this.movieRepository.update(movieModel);
+    public MovieServiceResponse update(MovieModel movieModel) {
+        MovieServiceResponse response = movieValidator.updateMovie(this.movieRepository.findById(movieModel.getId()));
+        if(response.getHttpStatus() == HttpStatus.OK){
+            this.movieRepository.update(movieModel);
+        }
+        return response;
     }
 }
